requirements.txt
pandas
numpy
sqlalchemy
psycopg2-binary
pyxirr
python-dotenv
python-dateutil
langchain
torch
torchvision
torchaudio
transformers
accelerate
ipykernel
sentence-transformers
pypdf
pydantic
huggingface-hub
spyder-kernels
matplotlib
numpy_financial
open-webui
llama-index-core
llama-index-llms-ollama
llama-index-embeddings-ollama
llama-index-vector-stores-postgres
llama-index-readers-file
pymupdf
tabulate
llama-index
langchain-ollama
langchainhub
